+++ packages/react-magma-dom/src/components/Dropdown/DropdownButton.tsx
@@ -2,6 +2,8 @@ import * as React from 'react';
 
 import { IconButton, ButtonIconPosition } from '../IconButton';
 import { CaretDownIcon } from '../Icon/types/CaretDownIcon';
+import { CaretLeftIcon } from '../Icon/types/CaretLeftIcon';
+import { CaretRightIcon } from '../Icon/types/CaretRightIcon';
 import { CaretUpIcon } from '../Icon/types/CaretUpIcon';
 import { DropdownContext, DropdownDropDirection } from '.';
 import { IconProps } from '../Icon/utils';
@@ -35,12 +37,21 @@ export const DropdownButton: React.FunctionComponent<DropdownButtonProps> = (
 ) => {
   const context = React.useContext(DropdownContext);
 
-  const buttonIcon =
-    context.dropDirection === DropdownDropDirection.up ? (
-      <CaretUpIcon size={10} testId="caretUp" />
-    ) : (
-      <CaretDownIcon size={10} testId="caretDown" />
-    );
+  function getButtonIcon(dropDirection: DropdownDropDirection) {
+    switch (dropDirection) {
+      case DropdownDropDirection.left:
+        return <CaretLeftIcon size={10} testId="caretLeft" />;
+      case DropdownDropDirection.right:
+        return <CaretRightIcon size={10} testId="caretRight" />;
+      case DropdownDropDirection.up:
+        return <CaretUpIcon size={10} testId="caretUp" />;
+
+      default:
+        return <CaretDownIcon size={10} testId="caretDown" />;
+    }
+  }
+
+  const buttonIcon = getButtonIcon(context.dropDirection);
 
   let children;
   const { icon = buttonIcon, ...other } = props;
@@ -63,7 +74,11 @@ export const DropdownButton: React.FunctionComponent<DropdownButtonProps> = (
       aria-expanded={context.isOpen}
       aria-haspopup="true"
       icon={icon}
-      iconPosition={ButtonIconPosition.right}
+      iconPosition={
+        context.dropDirection === DropdownDropDirection.left
+          ? ButtonIconPosition.left
+          : ButtonIconPosition.right
+      }
       onClick={handleClick}
       ref={context.toggleRef}
     >